+++ lib/sidekiq/api.rb
@@ -424,9 +424,10 @@ module Sidekiq
       Sidekiq.redis do |conn|
         workers = conn.smembers("workers")
         workers.each do |w|
-          msg, time = conn.mget("worker:#{w}", "worker:#{w}:started")
-          next unless msg
-          block.call(w, Sidekiq.load_json(msg), time)
+          json = conn.get("worker:#{w}")
+          next unless json
+          msg = Sidekiq.load_json(json)
+          block.call(w, msg, Time.at(msg['run_at']).to_s)
         end
       end
     end
@@ -435,6 +436,36 @@ module Sidekiq
       Sidekiq.redis do |conn|
         conn.scard("workers")
       end.to_i
+    end
+
+    # Prune old worker entries from the Busy set.  Worker entries
+    # can be orphaned if Sidekiq hard crashes while processing jobs.
+    # Default is to delete worker entries older than one hour.
+    #
+    # Returns the number of records removed.
+    def prune(older_than=60*60)
+      to_rem = []
+      Sidekiq.redis do |conn|
+        conn.smembers('workers').each do |w|
+          msg = conn.get("worker:#{w}")
+          if !msg
+            to_rem << w
+          else
+            m = Sidekiq.load_json(msg)
+            run_at = Time.at(m['run_at'])
+            # prune jobs older than one hour
+            if run_at < (Time.now - older_than)
+              to_rem << w
+            else
+            end
+          end
+        end
+      end
+
+      if to_rem.size > 0
+        Sidekiq.redis { |conn| conn.srem('workers', to_rem) }
+      end
+      to_rem.size
     end
   end