+++ Changes.md
@@ -2,6 +2,7 @@
 -----------
 
 - Auto-prune jobs older than one hour from the Workers page [#1508]
+- Add Sidekiq::Workers#prune which can perform the auto-pruning.
 - Fix issue where a job could be lost when an exception occurs updating
   Redis stats before the job executes [#1511]